Creating MainScreen style file and correcting names of files
